requirements.txt
absl-py==1.1.0
attrs==21.4.0
asyncio==3.4.3
click==8.1.3
cycler==0.11.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
fonttools==4.34.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
mediapipe==0.8.10.1
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
psycopg2-binary==2.9.3
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.39
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
